Delete no use code

+++ src/SparseGraph.h
@@ -501,7 +501,7 @@ void SparseGraph<node_type, edge_type>::RemoveEdge(int src, int dst){
 template <class node_type, class edge_type>
 void SparseGraph<node_type, edge_type>::SetEdgeCost(int src, int dst, double cost){
   
-  assert((src < m_nodes.size()) && (dst < m_nodes.size()) &&
+  assert(((unsigned int)src < m_nodes.size()) && ((unsigned int)dst < m_nodes.size()) &&
 	 "<SparseGraph::SetEdgeCost>: invalid index");
   
   for(EdgeList::iterator curEdge = m_edges[src].begin();
